AWQ GEMM backpropagation
