+++ awq/modules/linear.py
@@ -1,7 +1,10 @@
 import math
 import torch
 import torch.nn as nn
+from typing import Any, Tuple
+from torch.autograd.function import once_differentiable
 import awq_inference_engine  # with CUDA kernels
+from peft.tuners.lora import LoraLayer
 
 
 def make_divisible(c, divisor):
@@ -21,6 +24,35 @@ def calculate_zeros_width(in_features, group_size=128, pack_num=8):
     base_width = make_divisible(base_width, size_multiplier) * size_multiplier
     return base_width
 
+class WQLinear_GEMM_Propagator(torch.autograd.Function):
+    @staticmethod
+    def forward(input: torch.Tensor, qweight: torch.Tensor, scales: torch.Tensor, qzeros: torch.Tensor, split_k_iters: int) -> Any:
+        out = awq_inference_engine.gemm_forward_cuda(input.reshape(-1, input.shape[-1]), qweight, scales, qzeros, split_k_iters)
+        return out
+    
+    @staticmethod
+    def setup_context(ctx: Any, inputs: Tuple[Any], output: Any) -> Any:
+        qweight = inputs[1]
+        scales = inputs[2]
+        qzeros = inputs[3]
+        split_k_iters = inputs[4]
+        ctx.save_for_backward(qweight, scales, qzeros)
+        ctx.split_k_iters = split_k_iters
+    
+    @staticmethod
+    @once_differentiable
+    def backward(ctx, grad_output):
+        qweight, scales, qzeros = ctx.saved_tensors
+        split_k_iters = ctx.split_k_iters
+
+        # ! Hack: Don't train if the shapes of gradients and Q are different
+        if grad_output.shape[1] != qweight.shape[0]:
+            return torch.zeros(grad_output.shape[0], qweight.shape[0], device=qweight.device), None, None, None, None
+
+        backprop = awq_inference_engine.gemm_backward_cuda(grad_output.reshape(-1, grad_output.shape[-1]), qweight, scales, qzeros, split_k_iters)
+        return backprop, None, None, None, None
+
+
 class WQLinear_GEMM(nn.Module):
     def __init__(self, w_bit, group_size, in_features, out_features, bias, dev):
         super().__init__()
@@ -34,8 +66,9 @@ class WQLinear_GEMM(nn.Module):
         self.group_size = group_size if group_size != -1 else in_features
         
         # quick sanity check (make sure aligment)
-        assert self.in_features % self.group_size == 0
-        assert out_features % (32 // self.w_bit) == 0
+        if not isinstance(self, LoraLayer):
+            assert self.in_features % self.group_size == 0
+            assert out_features % (32 // self.w_bit) == 0
 
         self.register_buffer('qweight', torch.zeros((in_features, out_features // (32 // self.w_bit)), dtype=torch.int32, device=dev))
         self.register_buffer('qzeros', torch.zeros((in_features // self.group_size, out_features // (32 // self.w_bit)), dtype=torch.int32, device=dev))
@@ -94,19 +127,19 @@ class WQLinear_GEMM(nn.Module):
         
         return awq_linear
 
-    @torch.no_grad()
+    # @torch.no_grad()
     def forward(self, x):
         out_shape = x.shape[:-1] + (self.out_features, )
 
         input_dtype = x.dtype
         if input_dtype != torch.float16:
             x = x.half()
-        
-        out = awq_inference_engine.gemm_forward_cuda(x.reshape(-1, x.shape[-1]), self.qweight, self.scales, self.qzeros, 8)
-        
+
+        out = WQLinear_GEMM_Propagator.apply(x.reshape(-1, x.shape[-1]), self.qweight, self.scales, self.qzeros, 8)
+
         if input_dtype != torch.float16:
             out = out.to(dtype=input_dtype)
-        
+
         out = out + self.bias if self.bias is not None else out
         return out.reshape(out_shape)
     
@@ -204,18 +237,11 @@ class WQLinear_GEMV(nn.Module):
     def forward(self, x):
         out_shape = x.shape[:-1] + (self.out_features, )
         inputs = x.reshape(-1, x.shape[-1])
-
-        input_dtype = inputs.dtype
-        if input_dtype != torch.float16:
-            inputs = inputs.half()
         
         if inputs.shape[0] > 8:
             out = awq_inference_engine.gemmv2_forward_cuda(inputs, self.qweight, self.scales, self.qzeros, self.group_size, self.split_k_iters)
         else:
             out = awq_inference_engine.gemv_forward_cuda(inputs, self.qweight, self.scales, self.qzeros, self.group_size)
-
-        if input_dtype != torch.float16:
-            out = out.to(dtype=input_dtype)
         
         out = out + self.bias if self.bias is not None else out
         return out.reshape(out_shape)
